#40 Fix fromInfinite error
In the implementation of fromInfinite, we pass an infinite list to
Data.Primitive.Array.fromListN (bits + 1), but fromListN expects a list
a list of length (bits + 1), which results in a runtime error.

To fix this, we add a check in go to break recursion and to ensure that
the list that is passed into fromListN has length (bits + 1).